+++ R/utils.R
@@ -116,33 +116,7 @@ calculate_covariance_list <- function(psyList, lambdaList) {
 #' @noRd
 #' @noRd
 update_latent_scores <- function(X, thetaYList, ZOneDim, clusInd, qVec) {
-  n <- ncol(X)
-  active <- which(clusInd == 1)
-  Zmat <- get_z_mat_r(ZOneDim, length(qVec), n)
-  Y <- thetaYList@Y
-
-  for (k in active) {
-    lambda <- thetaYList@lambda[[k]]
-    psy <- thetaYList@psy[[k]]
-    qk <- qVec[k]
-    D <- t(lambda) %*% solve(psy + lambda %*% t(lambda), tol = 1e-100)
-    Sigma <- diag(qk) - D %*% lambda
-    Y[[k]] <- matrix(NA_real_, qk, n)
-
-    for (i in seq_len(n)) {
-      if (Zmat[k, i] == 0) {
-        Y[[k]][, i] <- mvtnorm::rmvnorm(1, mean = rep(0, qk), sigma = diag(qk))
-      } else {
-        Y[[k]][, i] <- mvtnorm::rmvnorm(
-          1,
-          mean = D %*% t(X[, i] - thetaYList@M[[k]]),
-          sigma = Sigma
-        )
-      }
-    }
-  }
-
-  Y
+  update_latent_scores_cpp(X, thetaYList, ZOneDim, clusInd, qVec)
 }